spam_checker_app/requirements.txt
cors==1.0.1
Django==4.2.4
django-cors-headers==4.2.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
